🚚 Move all files into one dir
