+++ static/app/components/externalIssues/ticketRuleModal.spec.tsx
@@ -6,10 +6,10 @@ import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
 import selectEvent from 'sentry-test/selectEvent';
 
 import {addSuccessMessage} from 'sentry/actionCreators/indicator';
+import TicketRuleModal from 'sentry/components/externalIssues/ticketRuleModal';
 import {makeCloseButton} from 'sentry/components/globalModal/components';
 import type {IssueAlertRuleAction} from 'sentry/types/alerts';
 import type {IssueConfigField} from 'sentry/types/integrations';
-import TicketRuleModal from 'sentry/views/alerts/rules/issue/ticketRuleModal';
 
 jest.unmock('sentry/utils/recreateRoute');
 jest.mock('sentry/actionCreators/indicator');
