+++ static/app/components/externalIssues/ticketRuleModal.tsx
